requirements.txt
streamlit
google-generativeai
Pillow
pandas
plotly
fpdf2
